Changed package names, change CityNotFoundException to EntityNotFoundException so it can be reused for other еentities which happen to not be found during CRUD operation.
