--- src/main/java/com/cifolio/cifolio/dto/city/CityDto.java
@@ -1,11 +1,0 @@
-package com.cifolio.cifolio.dto.city;
-
-import lombok.*;
-
-@AllArgsConstructor
-@Getter
-public class CityDto {
-    private Long id;
-    private String name;
-    private String photo;
-}
