--- src/main/java/com/cifolio/cifolio/dto/user/UserDto.java
@@ -1,12 +1,0 @@
-package com.cifolio.cifolio.dto.user;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserDto {
-    private String username;
-    private String password;
-    private String email;
-}
