+++ src/main/java/com/cifolio/cifolio/dtos/city/CityDto.java
@@ -1,0 +1,11 @@
+package com.cifolio.cifolio.dtos.city;
+
+import lombok.*;
+
+@AllArgsConstructor
+@Getter
+public class CityDto {
+    private Long id;
+    private String name;
+    private String photo;
+}
