--- src/main/java/com/cifolio/cifolio/exception_handling/ApiError.java
@@ -1,15 +1,0 @@
-package com.cifolio.cifolio.exception_handling;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@Getter
-public class ApiError {
-    private String message;
-    private HttpStatus status;
-    private LocalDateTime timestamp;
-}
