--- src/main/java/com/cifolio/cifolio/exception_handling/exceptions/CityNotFoundException.java
@@ -1,7 +1,0 @@
-package com.cifolio.cifolio.exception_handling.exceptions;
-
-public class CityNotFoundException extends RuntimeException {
-    public CityNotFoundException(String message) {
-        super(message);
-    }
-}
