+++ src/main/java/com/cifolio/cifolio/exception_handling/exceptions/EntityNotFoundException.java
@@ -1,0 +1,7 @@
+package com.cifolio.cifolio.exception_handling.exceptions;
+
+public class EntityNotFoundException extends RuntimeException {
+    public EntityNotFoundException(String message) {
+        super(message);
+    }
+}
