+++ src/main/java/com/cifolio/cifolio/filters/AuthenticationFilter.java
@@ -1,0 +1,59 @@
+package com.cifolio.cifolio.filters;
+
+import com.cifolio.cifolio.service.token.TokenService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+
+import static com.cifolio.cifolio.constants.SecurityConstants.JWT_ACCESS_TOKEN_NAME;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+
+@RequiredArgsConstructor
+public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
+    private final AuthenticationManager authenticationManager;
+    private final TokenService tokenService;
+
+    @Override
+    public Authentication attemptAuthentication(HttpServletRequest request, HttpServletResponse response) throws AuthenticationException {
+        try {
+            Map<String, String> requestMap = new ObjectMapper().readValue(request.getInputStream(), Map.class);
+            String username = requestMap.get("username");
+            String password = requestMap.get("password");
+
+            UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(username, password);
+            return authenticationManager.authenticate(authenticationToken);
+        } catch(IOException exception) {
+            throw new RuntimeException(exception);
+        }
+    }
+
+    @Override
+    protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authentication) throws IOException, ServletException {
+        String accessToken = tokenService.generateToken(authentication);
+        Cookie userAuthenticationCookie = new Cookie(JWT_ACCESS_TOKEN_NAME, accessToken);
+
+        userAuthenticationCookie.setHttpOnly(true);
+        userAuthenticationCookie.setMaxAge(3600 * 24);
+        response.addCookie(userAuthenticationCookie);
+
+        Map<String, String> responseBody = new HashMap<>();
+        responseBody.put("username", authentication.getName());
+        responseBody.put("role", authentication.getAuthorities().stream().findFirst().get().toString());
+
+        response.setContentType(APPLICATION_JSON_VALUE);
+        new ObjectMapper().writeValue(response.getOutputStream(), responseBody);
+    }
+}
