+++ src/main/java/com/cifolio/cifolio/repositories/UserRepository.java
@@ -1,0 +1,12 @@
+package com.cifolio.cifolio.repositories;
+
+import com.cifolio.cifolio.models.user.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+}
