--- src/main/java/com/cifolio/cifolio/repository/TravelGuideRepository.java
@@ -1,9 +1,0 @@
-package com.cifolio.cifolio.repository;
-
-import com.cifolio.cifolio.model.city.TravelGuide;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TravelGuideRepository extends JpaRepository<TravelGuide, Long> {
-}
