--- src/main/java/com/cifolio/cifolio/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.cifolio.cifolio.repository;
-
-import com.cifolio.cifolio.model.user.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-}
