+++ src/main/java/com/cifolio/cifolio/service/user/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.cifolio.cifolio.service.user;
 
-import com.cifolio.cifolio.model.user.AuthenticatedUser;
-import com.cifolio.cifolio.repository.UserRepository;
+import com.cifolio.cifolio.models.user.AuthenticatedUser;
+import com.cifolio.cifolio.repositories.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
