+++ src/main/java/com/cifolio/cifolio/service/user/UserService.java
@@ -1,7 +1,7 @@
 package com.cifolio.cifolio.service.user;
 
-import com.cifolio.cifolio.model.user.User;
-import com.cifolio.cifolio.repository.UserRepository;
+import com.cifolio.cifolio.models.user.User;
+import com.cifolio.cifolio.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
